% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLabel.R
\name{colInfo}
\alias{colInfo}
\title{Get Details for All Columns}
\usage{
colInfo(df, limit = 0, rep = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{limit}{The maximum number of labels per column (default: 0, which means all labels).}

\item{rep}{A logical value indicating whether to allow repetitions (default: FALSE).}
}
\value{
A data frame with variable names, labels, category values, and category labels.
}
\description{
Retrieves labels and details for all columns in the data frame.
}
\examples{
colInfo(df, limit = 1)
}
