% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLabel.R
\name{listVars}
\alias{listVars}
\title{Get Variable Labels}
\usage{
listVars(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with variable names and their corresponding labels.
}
\description{
Retrieves all variable names and their labels from the data frame.
}
\examples{
listVars(df)
}
