% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLabel.R
\name{setLbl}
\alias{setLbl}
\title{Set Label for a Column}
\usage{
setLbl(df, col, suffix = "_label")
}
\arguments{
\item{df}{A data frame.}

\item{col}{The column name for which labels are to be set.}

\item{suffix}{A string to append to the label column name (default: "_label").}
}
\value{
A data frame with the labels merged into the original data frame.
}
\description{
Merges labels into the original data frame for a specified column.
}
\examples{
setLbl(df, "V5", suffix = "_label")
}
