% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLabel.R
\name{setlstLbl}
\alias{setlstLbl}
\title{Set Labels for All Columns by Selected Names}
\usage{
setlstLbl(df, columns = c())
}
\arguments{
\item{df}{A data frame.}

\item{columns}{A character vector of column names for which labels should be set.}
}
\value{
A data frame with labels set for the specified columns.
}
\description{
Sets labels for all columns in a data frame that match a specified list of column names.
}
\examples{
setlstLbl(df, columns = c("V1", "V2"))
}
