% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabel.R
\name{getLbl}
\alias{getLbl}
\title{Get Label for a Column}
\usage{
getLbl(df, col, suffix = "_labels", limit = 0)
}
\arguments{
\item{df}{A data frame containing the column.}

\item{col}{The column name for which labels are to be retrieved.}

\item{suffix}{A string to append to the label column name (default: "_labels").}

\item{limit}{The maximum number of labels to return (default: 0, which means all labels).}
}
\value{
A data frame with the column values and their corresponding labels.
}
\description{
Retrieves the labels of a specified column in a data frame.
}
\details{
The function retrieves the \code{labels} attribute of a factor variable and uses it to create a new column with descriptive labels. The new column name is either specified by the \code{bycol} parameter or dynamically generated using the \code{suffix} parameter.
}
\examples{
# Create a sample data frame with a factor column
# Load the package
library(defactor)

#Load the required package
library(data.table)
# Create the data table
df <- data.table(
  C_COW_ALPHA = c("BOS", "BOS", "BOS", "BOS", "BOS"),
  COW = c(346, 346, 346, 346, 346),
  B_COUNTRY_ALPHA = c("BIH", "BIH", "BIH", "BIH", "BIH"),
  V2 = c(70, 70, 70, 70, 70),
  V2A = c(70, 70, 70, 70, 70),
  V3 = c(1, 2, 3, 4, 5),
  V4 = c(1, 1, 1, 1, 1),
  V5 = c(1, 1, 1, 1, 1),
  V6 = c(1, 2, 2, 1, 2)
)

# Add column labels as attributes
attr(df$COW, "label") <- "colum1"
attr(df$V2, "label") <- "colum2"
attr(df$V2A, "label") <- "colum3"
attr(df$V4, "label") <- "colum4"
attr(df$V5, "label") <- "colum5"
attr(df$V6, "label") <- "colum6"

# Add value labels as attributes
attr(df$COW, "labels") <- c("Bosnia and Herzegovina" = 346)
attr(df$V2, "labels") <- c("Bosnia Herzegovina" = 70)
attr(df$V2A, "labels") <- c("Bosnia Herzegovina" = 70)
attr(df$V4, "labels") <- c("Very important" = 1)
attr(df$V5, "labels") <- c("Very important" = 1)
attr(df$V6, "labels") <- c("Very important" = 1, "Rather important" = 2)
attr(df$V6, "labels") <- c("Very important" = 1, "Rather important" = 2)
}
